Level One: Grounded Created
I created the ground/hills for the first level, the sky still needs to
be added in, however I want your input on the ground before adding that
in. After that I will be adding in the obstacle layer. You will need to
open it in Tiled in order to view the map.