Code to Create Tilemap
There are three lines that create the timely in gamedriver, Dr. Palmer
took a look at them and said they are all correct, however our game
still crashes so there is another issue. Heres the code so everyone has
it, if you are trying to do something and need it to not crash just
comment out line 144, 145.